refactor
- userId Long -> Integer 변경
- userId @PathVariable 방식으로 임시 설정
- DTO @JsonNaming 추가

+++ src/main/java/com/bovo/Bovo/modules/archive/service/MyBooksServiceImpl.java
@@ -21,7 +21,7 @@ public class MyBooksServiceImpl implements MyBooksService {
     }
 
     @Override
-    public MyBooksResponse getMyBooks(Long userId) {
+    public MyBooksResponse getMyBooks(Integer userId) {
         // 사용자 ID에 해당하는 내 서재의 책 목록을 조회
         List<MyBooks> myBooksList = myBooksRepository.findByUserId(userId);
